GET PhotoResult.json Successfully

+++ app/src/main/java/cn/com/wosuo/taskrecorder/vo/UserLocation.java
@@ -1,23 +1,23 @@
 package cn.com.wosuo.taskrecorder.vo;
 
 public class UserLocation {
-    private int positionX;
-    private int positionY;
+    private double positionX;
+    private double positionY;
     private int coordinate;
 
-    public int getPositionX() {
+    public double getPositionX() {
         return positionX;
     }
 
-    public void setPositionX(int positionX) {
+    public void setPositionX(double positionX) {
         this.positionX = positionX;
     }
 
-    public int getPositionY() {
+    public double getPositionY() {
         return positionY;
     }
 
-    public void setPositionY(int positionY) {
+    public void setPositionY(double positionY) {
         this.positionY = positionY;
     }
